Ref: Separate data integration code by purpose
This commit separates the codebase into distinct files based on their responsibilities. There are now separate files for the handbook reader, visualizer, Unit model, and startup script. This should make the code easier to navigate and maintain in the long term.
