+++ .gitignore
@@ -1,4 +1,4 @@
-DataIntegration/handbooks/*
+Backend/DataIntegration/handbooks/*
 
 #######################
 ##### Python PyCharm #####
